refactor(ui): ensure globals file is named `*_globals.hpp`
Some libraries had globals as a singular.
