+++ ui/memory/memory_globals.hpp
